requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
autocommand==2.2.2
certifi==2025.6.15
chardet==5.2.0
charset-normalizer==3.4.2
dotenv==0.9.9
idna==3.10
jaraco.context==6.0.1
jaraco.functools==4.2.1
jaraco.text==4.0.0
macholib==1.16.3
modulegraph==0.19.6
more-itertools==10.7.0
packaging==25.0
py2app==0.28.8
pydantic_core==2.33.2
python-dotenv==1.1.1
requests==2.32.4
setuptools==80.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
